Move TimeControllerClock into separate file
