+++ rosbag2_cpp/include/rosbag2_cpp/clocks/player_clock.hpp
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef ROSBAG2_CPP__PLAYER_CLOCK_HPP_
-#define ROSBAG2_CPP__PLAYER_CLOCK_HPP_
+#ifndef ROSBAG2_CPP__CLOCKS__PLAYER_CLOCK_HPP_
+#define ROSBAG2_CPP__CLOCKS__PLAYER_CLOCK_HPP_
 
 #include <chrono>
 #include <functional>
@@ -69,60 +69,6 @@ public:
   virtual double get_rate() const = 0;
 };
 
-/**
- * Version of the PlayerClock interface that has control over time.
- * It does not listen to any external ROS Time Source and can optionally publish /clock
- */
-class TimeControllerClockImpl;
-class TimeControllerClock : public PlayerClock
-{
-public:
-  /**
-   * Constructor.
-   *
-   * \param starting_time: provides an initial offset for managing time
-   *    This will likely be the timestamp of the first message in the bag
-   * \param rate: Rate of playback, a unit-free ratio. 1.0 is real-time
-   * \param now_fn: Function used to get the current steady time
-   *   defaults to std::chrono::steady_clock::now
-   *   Used to control for unit testing, or for specialized needs
-   * \throws std::runtime_error if rate is <= 0
-   */
-  ROSBAG2_CPP_PUBLIC
-  TimeControllerClock(
-    ROSTimePoint starting_time,
-    double rate = 1.0,
-    NowFunction now_fn = std::chrono::steady_clock::now);
-
-  ROSBAG2_CPP_PUBLIC
-  virtual ~TimeControllerClock();
-
-  /**
-   * Calculate and return current ROSTimePoint based on starting time, playback rate, pause state.
-   */
-  ROSBAG2_CPP_PUBLIC
-  ROSTimePoint now() const override;
-
-  /**
-   * Try to sleep (non-busy) the current thread until the provided time is reached - according to this Clock
-   *
-   * Return true if the time has been reached, false if it was not successfully reached after sleeping
-   * for the appropriate duration.
-   * The user should not take action based on this sleep until it returns true.
-   */
-  ROSBAG2_CPP_PUBLIC
-  bool sleep_until(ROSTimePoint until) override;
-
-  /**
-   * Return the current playback rate.
-   */
-  ROSBAG2_CPP_PUBLIC
-  double get_rate() const override;
-
-private:
-  std::unique_ptr<TimeControllerClockImpl> impl_;
-};
-
 }  // namespace rosbag2_cpp
 
-#endif  // ROSBAG2_CPP__PLAYER_CLOCK_HPP_
+#endif  // ROSBAG2_CPP__CLOCKS__PLAYER_CLOCK_HPP_
