+++ rosbag2_cpp/src/rosbag2_cpp/clocks/time_controller_clock.cpp
@@ -17,7 +17,7 @@
 #include <thread>
 
 #include "rcpputils/thread_safety_annotations.hpp"
-#include "rosbag2_cpp/player_clock.hpp"
+#include "rosbag2_cpp/clocks/time_controller_clock.hpp"
 #include "rosbag2_cpp/types.hpp"
 
 namespace rosbag2_cpp
